+++ CHANGELOG.md
@@ -8,6 +8,11 @@
 
 * A huge change of script structure, it is not compatible with previous versions.
 
-## `0.2.4`/`0.2.5`/`0.2.6` at `2021-09-24`
+## `0.2.4`/`0.2.5`/`0.2.6`/`0.2.7` at `2021-09-24`
 
 * Updating for bug fixes.
+
+## `0.2.8` at `2021-09-25`
+
+* A select input can reflect values to another with `reflect()` method and clean all items with `cleanItems()` method;
+* `get()` method has type in `PglyBaseFormEngine` and `PglyGroupFormComponent`.